// SiscoV3/src/app/proveedor/enums.ts
export enum BUTTONS_TOOLBAR {
    EDITAR = 'Editar',
    AGREGAR = 'Agregar',
    ELIMINAR = 'Eliminar',
    AGREGAR_ENTIDAD = 'Agregar Sucursal',
    AGREGAR_DIRECCION = 'Agregar Dirección',
    SELECCION_DE_TALLER='Seleccionar'
}

export enum ITEM_STORAGE {
    ITEM_DATOS_PROVEEDOR = 'datosProveedor'
}